third_party/blink/web_tests/platform/generic/fast/dom/HTMLImageElement/parse-src-expected.txt

Test that when using the JavaScript bindings to get the value of a URL attribute such as 'src', leading whitespace is ignored

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testURL("src", url) is url
PASS testURL("src", "\n" + url) is url
PASS testURL("src", " " + url) is url
PASS testURL("lowsrc", url) is url
PASS testURL("lowsrc", "\n" + url) is url
PASS testURL("lowsrc", " " + url) is url
PASS testURL("longDesc", url) is url
PASS testURL("longDesc", "\n" + url) is url
PASS testURL("longDesc", " " + url) is url
PASS successfullyParsed is true

TEST COMPLETE
